feat: avoid refresh token multiple calls and move table utils to table component
