+++ packages/react-data-provider/package.json
@@ -14,6 +14,7 @@
     "react": "^18.2.0"
   },
   "dependencies": {
-    "axios": "^0.24.0"
+    "axios": "^0.24.0",
+    "mem": "^10.0.0"
   }
 }
