--- packages/react-material-ui/src/utils/index.ts
@@ -1,3 +1,0 @@
-import createTableStyles from './table/createTableStyles';
-
-export { createTableStyles };
